+++ R/AcqFunctionEHVIGH.R
@@ -116,6 +116,7 @@ AcqFunctionEHVIGH = R6Class(
 
     #' @description
     #' Update the acquisition function and set `ys_front`, `ref_point`, `hypervolume` and `gh_data`.
+    #' See [AcqFunction] for a general description of `$update()`.
     update = function() {
       n_obj = length(self$archive$cols_y)
       ys = self$archive$data[, self$archive$cols_y, with = FALSE]